Move the Submenu class inside migration/
Most of the migration classes
will be in this directory.
